refactor: cleanup slide action utils
